Add patch to increase ZoneHigh memory area
When qemu runs with a large number of maxcpus (64) and also a large number of
DIMMs (e.g. 255), seabios runs out of memory. This results in memory hotplug
errors, since the memory DIMM objects fail to be allocated.

This will also be needed to describe large number of memory DIMMs (1024) in the
future.